+++ parsec/algorithm/aead.go
