+++ parsec/algorithm/cipher.go
