+++ parsec/algorithm/hash.go
