+++ parsec/algorithm/mac.go
