+++ parsec/algorithm/test/hash_test.go
